Adopt redux-router for navigation control
Also, in the process:

- move the theming stuff out of App so that testing won't need it
- add routing reducers
- move store creation logic into store.js
