+++ package.json
@@ -64,6 +64,7 @@
     "react-quill": "^1.0.0-beta-1",
     "react-redux": "^5.0.3",
     "react-router": "^3.0.2",
+    "react-router-redux": "^4.0.8",
     "react-s3-uploader": "^4.0.3",
     "react-tap-event-plugin": "^2.0.1",
     "reactstrap": "^4.2.0",
