+++ src/App.js
@@ -1,25 +1,14 @@
-import React, { PropTypes, Component } from "react"; // eslint-disable-line no-unused-vars
-import MuiThemeProvider from "material-ui/styles/MuiThemeProvider";
-import { IntlProvider, addLocaleData } from "react-intl";
-import getMuiTheme from "material-ui/styles/getMuiTheme";
+import React from "react"; // eslint-disable-line
 import routeConfig from "./routes";
 import { Router, browserHistory } from "react-router";
-import locales from "../public/locales.json";
-import "bootstrap/dist/css/bootstrap.min.css";
-
-// This replaces the textColor value on the palette
-// and then update the keys for each component that depends on it.
-// More on Colors: http://www.material-ui.com/#/customization/colors
-const muiTheme = getMuiTheme({
-  palette: {
-    primary1Color: "#3f51b2"
-  },
-  appBar: {}
-});
-
+import { IntlProvider, addLocaleData } from "react-intl";
 import en from "react-intl/locale-data/en";
 import fr from "react-intl/locale-data/fr";
 import es from "react-intl/locale-data/es";
+import locales from "../public/locales.json";
+import { syncHistoryWithStore } from "react-router-redux";
+import store from "./store";
+import "bootstrap/dist/css/bootstrap.min.css";
 
 addLocaleData([...en, ...fr, ...es]);
 
@@ -34,18 +23,18 @@ function createElement(Component, props) {
   );
 }
 
-class App extends Component {
+const history = syncHistoryWithStore(browserHistory, store);
+
+export class App extends React.Component {
   render() {
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
-        <Router
-          createElement={createElement}
-          history={browserHistory}
-          routes={routeConfig}
-        />
-      </MuiThemeProvider>
+      <Router
+        createElement={createElement}
+        history={history}
+        routes={routeConfig}
+      />
     );
   }
 }
 
-export default App;
+module.exports = { App, store };
