+++ src/routes.js
@@ -15,7 +15,6 @@ import Case from "./containers/Case/Case";
 import Organization from "./containers/Organization";
 import Method from "./containers/Method";
 import Add from "./components/Add/Add";
-// import EditCase from "./containers/EditCase";
 import CaseEditorContainer from "./containers/CaseEditorContainer";
 import QuickSubmitPicker
   from "./components/QuickSubmitPicker/QuickSubmitPicker";
