+++ src/store.js
@@ -1,8 +1,11 @@
-import { compose, createStore, applyMiddleware } from "redux";
 import * as storage from "redux-storage";
+import { compose, createStore, applyMiddleware } from "redux";
 import createEngine from "redux-storage-engine-localstorage";
 const engine = createEngine("participedia");
 import thunkMiddleware from "redux-thunk";
+import { routerMiddleware } from "react-router-redux";
+import { browserHistory } from "react-router";
+const routeMiddleware = routerMiddleware(browserHistory);
 import rootReducer from "./reducers";
 
 function configureStore() {
@@ -12,7 +15,7 @@ function configureStore() {
     rootReducer,
     initialState,
     compose(
-      applyMiddleware(thunkMiddleware, storageMiddleware),
+      applyMiddleware(thunkMiddleware, storageMiddleware, routeMiddleware),
       window.devToolsExtension ? window.devToolsExtension() : f => f
     )
   );
@@ -20,4 +23,6 @@ function configureStore() {
   return store;
 }
 
-export default configureStore;
+let store = configureStore();
+
+export default store;
